refactor(homePage): delete files and modify components
