--- components/Welcome/Welcome.module.css
@@ -1,345 +1,0 @@
-/* ------ WRAPPER OF HEADER - MAIN ------ */
-.wrapperContainer{
-    min-height: 100dvh;
-    display: grid;
-    grid-template-columns: 9rem 1fr;
-}
-
-/* ------ HERO SECTION ------ */
-.heroSection{
-    min-height: 100dvh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 1rem 3rem;
-    gap: 2rem;
-}
-
-.heroSection__titles{
-    background: linear-gradient(360deg, #3339, transparent 100%);
-    width: 35em;
-    padding: 2rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    gap: 2rem;
-}
-
-.heroSection__titles h1{
-    font-size: 2.5em;
-}
-
-.heroSection__titles h2{
-    font-size: 2.25em;
-    opacity: .8;
-}
-
-.heroSection__carrousel{
-    border-radius: 2rem;
-    flex-grow: 2;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 500px;
-    height: 500px;
-}
-
-.heroSection__image{
-    border-radius: inherit;
-    height: 100%;
-    width: 100%;
-    object-fit: cover;
-}
-
-.heroSection__titles p{
-    text-align: center;
-    padding: 0 2rem;
-    color: #fff;
-    font-size: 0.90em;
-    text-wrap: balance;
-    font-weight: 300;
-    line-height: 1.75em;
-}
-
-/* ------ BANK RESOURCE SECTION ------ */
-.bankResources{
-    min-height: 100dvh;
-}
-
-.bankResources__title{
-    background: linear-gradient(180deg, rgba(0, 0, 0, .3), transparent 100%);
-    padding: 1rem;
-    font-size: 1.6rem;
-    text-align: center;
-    font-weight: 700;
-}
-
-.bankResources__items{
-    padding: 3rem;
-    display: grid;
-    gap: 2rem;
-    grid-template-columns: repeat(3, 1fr);
-    grid-template-rows: repeat(2, 1fr);
-}
-
-.bankResources__itemBox{
-    min-height: 12rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    background: #eeed;
-    color: #333;
-    padding: 1rem;
-    clip-path: polygon(0px 0px, 100% 0px, 100.00% 85%, 90% 100%, 0px 100%);
-    position: relative;
-}
-
-.bankResources__itemBox::before{
-    position: absolute;
-    top: 0;
-    left: 0;
-    content: "";
-    width: 100%;
-    height: 4px;
-    background: rgb(98, 26, 255);
-}
-
-.bankResources__itemBox div{
-    display: flex;
-    gap: .75rem;
-    align-items: center;
-}
-
-.bankResources__itemBox__icon{
-    box-sizing: content-box;
-    font-size: 1.3rem;
-    border: 1px solid #000;
-    border-radius: 50%;
-    padding: .25rem;
-}
-
-.bankResources__itemBox__title{
-    font-size: 1.4rem;
-}
-
-.bankResources__itemBox__description{
-    font-size: .9rem;
-    text-wrap: balance;
-}
-
-.bankResources__itemBox__moreInfoLink{
-    border-radius: 1rem;
-    padding: .25rem 1rem;
-    background: rgb(194, 63, 255);
-    width: max-content;
-    text-align: center;
-    display: flex;
-    align-items: center;
-    gap: .5rem;
-    transition: scale .15s ease-in-out;
-}
-
-.bankResources__itemBox__moreInfoLink:hover{
-    background-color: rgb(194, 63, 255, .8);
-    scale: 1.05;
-}
-
-/* ------ BANK ACTIONS SECTION ------ */
-.actionContainer{
-    background-color: rgba(0, 0, 0, .2);
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    grid-template-rows: repeat(3, 250px);
-}
-
-.actionContainer img:nth-child(3n){
-    grid-column: 2 / 3;
-}
-
-.actionContainer div:nth-child(4){
-    grid-column: 1 / 2;
-    grid-row: 2 / 3;
-    justify-self: end;
-}
-
-.actionContainer__img{
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-.actionContainer__content{
-    display: flex;
-    padding: 0 3rem;
-    align-items: center;
-}
-
-.actionContainer__content div{
-    width: 100%;
-    height: 70%;
-    display: grid;
-    grid-template-rows: 1fr 1fr 3fr;
-    align-items: center;
-}
-
-.actionContainer__content h4{
-    font-size: 1.5rem;
-    font-weight: 700
-}
-
-.actionContainer__content span{
-    font-size: .9rem;
-}
-
-.actionContainer__content a{
-    width: 8rem;
-    background-color: rgb(194, 63, 255);
-    text-align: center;
-    padding: .25rem 1rem;
-    border-radius: 2rem;
-    transition: scale .15s ease-in-out;
-}
-
-.actionContainer__content a:hover{
-    background-color: rgba(194, 63, 255, .8);
-    scale: 1.05;
-}
-
-/* ------ BANK FOOTER SECTION ------ */
-.footer{
-    padding-top: 2rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-between;
-    gap: 2rem;
-}
-
-.footer__wrapperInfo{
-    width: 80%;
-    display: grid;
-    grid-template-areas: 
-    "contact contact"
-    "form links"
-    ;
-    gap: 2rem;
-}
-
-.footerInfoContainer > h5{
-    padding: .75rem 0;
-    font-size: 1.1rem;
-    font-weight: 600;
-    position: relative;
-    margin-bottom: .75rem;
-}
-
-.footerInfoContainer > h5::after{
-    content: "";
-    position: absolute;
-    left: 0;
-    bottom: 0;
-    width: 100%;
-    height: 3px;
-    background: #3337;
-}
-
-.footerInfoContainer > p{
-    font-size: .75rem;
-    opacity: .9;
-    margin-bottom: .75rem;
-}
-
-.footer__form{
-    grid-area: form;
-}
-
-.footer__form form{
-    display: flex;
-    flex-wrap: wrap;
-}
-
-.footer__form :is(input, select),
-.footer__links a{
-    padding: .5rem 1rem;
-}
-
-.footer__form input[type="email"]{
-    flex-grow: 3;
-    flex-shrink: 3;
-    outline: none;
-    color: #333;
-}
-
-.footer select,
-.footer__links a{
-    background-color: #3339;
-    flex-grow: 2;
-}
-
-.footer__form input[type="submit"]{
-    background-color: #333d;
-    flex-grow: 1;
-    font-weight: 300;
-    cursor: pointer;
-    transition: background-color .15s ease-in-out;
-}
-
-.footer__form input[type="submit"]:hover{
-    background-color: #222;
-}
-
-.footer__contact{
-    grid-area: contact;
-}
-
-.footer__contact ul{
-    display: flex;
-    flex-wrap: wrap;
-    gap: 1rem;
-}
-.footer__contact div{
-    border: 1px solid #333;
-    border-radius: .25rem;
-    background-color: #3333;
-    font-size: .9rem;
-    display: flex;
-    flex-direction: column;
-    gap: .25rem;
-    flex: 1 1 15rem;
-    padding: 1rem;
-}
-
-.footer__contact div h5{
-    font-weight: 600;
-    text-transform: uppercase;
-}
-
-.footer__links{
-    grid-area: links;
-}
-
-.footer__logo{
-    width: 100%;
-    padding: .65rem 0;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-/* ------ MEDIA QUERIES ------ */
-@media (max-width: 820px){
-    .footer__wrapperInfo{
-        display: flex;
-        flex-direction: column;
-    }
-}
-
-/* @media (max-width: 600px){
-} */
-
-/* @media (max-width: 380px){
-} */
-
-/* @media (max-width: 280px){
-} */
