--- components/Welcome/bankActionsList.jsx
@@ -1,32 +1,0 @@
-import BANNER_1 from "@/public/banner_1.jpeg"
-import BANNER_2 from "@/public/banner_2.jpeg"
-import BANNER_3 from "@/public/banner_3.jpeg"
-
-const bankActionsList = [
-    {
-        title: "Turnos online",
-        description: "Solicitalo simple y rápido",
-        btn: "Pedilo ahora",
-        href: "#",
-        image: BANNER_1,
-        alt: "Una persona pidiendo un turno online con su telefono"
-    },
-    {
-        title: "Prevención de fraudes",
-        description: "Operá seguro en Canales Digitales",
-        btn: "Conocé más",
-        href: "#",
-        image: BANNER_2,
-        alt: "Imagen de un candado simbolizando una red segura"
-    },
-    {
-        title: "Información para personas usuarios de servicios financieros",
-        description: "Consultá lo que dispone el Banco Central",
-        btn: "Ingresa aquí",
-        href: "#",
-        image: BANNER_3,
-        alt: "Primera vista de un celular"
-    }
-]
-
-export default bankActionsList
