--- components/Welcome/bankContacts.js
@@ -1,32 +1,0 @@
-const bankContacts = [
-    {
-        "title": "Sucursal Central Plaza",
-        "number": "+54-11-1234-5678"
-    },
-    {
-        "title": "Sucursal Norte Vista",
-        "number": "+54-11-2345-6789"
-    },
-    {
-        "title": "Sucursal Sur Maravilla",
-        "number": "+54-11-3456-7890"
-    },
-    {
-        "title": "Sucursal Este Horizonte",
-        "number": "+54-11-4567-8901"
-    },
-    {
-        "title": "Sucursal Oeste Jardines",
-        "number": "+54-11-5678-9012"
-    },
-    {
-        "title": "Sucursal Villa Verde",
-        "number": "+54-11-6789-0123"
-    },
-    {
-        "title": "Sucursal Las Colinas",
-        "number": "+54-11-7890-1234"
-    }
-]
-
-export default bankContacts
