+++ components/autoImagesCarrousel/autoImagesCarrousel.module.css
@@ -1,0 +1,6 @@
+.heroSection__image {
+  border-radius: inherit;
+  height: 100%;
+  width: 100%;
+  object-fit: cover;
+}
