+++ components/footerHome/footerHome.module.css
@@ -1,0 +1,127 @@
+/* ------ BANK FOOTER SECTION ------ */
+.footer {
+  padding-top: 2rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  gap: 2rem;
+}
+
+.footer__wrapperInfo {
+  width: 80%;
+  display: grid;
+  grid-template-areas:
+    "contact contact"
+    "form links";
+  gap: 2rem;
+}
+
+.footerInfoContainer > h5 {
+  padding: 0.75rem 0;
+  font-size: 1.1rem;
+  font-weight: 600;
+  position: relative;
+  margin-bottom: 0.75rem;
+}
+
+.footerInfoContainer > h5::after {
+  content: "";
+  position: absolute;
+  left: 0;
+  bottom: 0;
+  width: 100%;
+  height: 3px;
+  background: #3337;
+}
+
+.footerInfoContainer > p {
+  font-size: 0.75rem;
+  opacity: 0.9;
+  margin-bottom: 0.75rem;
+}
+
+.footer__form {
+  grid-area: form;
+}
+
+.footer__form form {
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.footer__form :is(input, select),
+.footer__links a {
+  padding: 0.5rem 1rem;
+}
+
+.footer__form input[type="email"] {
+  flex-grow: 3;
+  flex-shrink: 3;
+  outline: none;
+  color: #333;
+}
+
+.footer select,
+.footer__links a {
+  background-color: #3339;
+  flex-grow: 2;
+}
+
+.footer__form input[type="submit"] {
+  background-color: #333d;
+  flex-grow: 1;
+  font-weight: 300;
+  cursor: pointer;
+  transition: background-color 0.15s ease-in-out;
+}
+
+.footer__form input[type="submit"]:hover {
+  background-color: #222;
+}
+
+.footer__contact {
+  grid-area: contact;
+}
+
+.footer__contact ul {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
+}
+.footer__contact div {
+  border: 1px solid #333;
+  border-radius: 0.25rem;
+  background-color: #3333;
+  font-size: 0.9rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.25rem;
+  flex: 1 1 15rem;
+  padding: 1rem;
+}
+
+.footer__contact div h5 {
+  font-weight: 600;
+  text-transform: uppercase;
+}
+
+.footer__links {
+  grid-area: links;
+}
+
+.footer__logo {
+  width: 100%;
+  padding: 0.65rem 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+/* ------ MEDIA QUERIES ------ */
+@media (max-width: 820px) {
+  .footer__wrapperInfo {
+    display: flex;
+    flex-direction: column;
+  }
+}
